qa_otus_sf/my_test_project/requirements.txt
pytest==5.0.1
selenium-3.141.0
requests==2.22.0
browsermob-proxy==0.8.0
pytest-html==2.0.0
